% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_xy.R
\name{classify_xy}
\alias{classify_xy}
\title{Detect spatial information in data columns}
\usage{
classify_xy(list, dim)
}
\arguments{
\item{list}{(list) A metajam list output for a single data entity.}

\item{dim}{(character) "x" or "y", dimension to look for.}
}
\value{
A logical vector classifying variables as TRUE, containing spatial information, or FALSE, not containing spatial information of the dimension specifed. Variables are in order as appear in data.
}
\description{
Detect spatial information in data columns
}
