% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapie_shiny.R
\name{datapie_shiny}
\alias{datapie_shiny}
\title{A data package interface for evaluation (datapie)}
\usage{
datapie_shiny()
}
\value{
A UI for visual exploration of a dataset
}
\description{
A data package interface for evaluation (datapie)
}
\examples{
# Open the UI
datapie_shiny()

}
