% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package_read.R
\name{guess_read_fnc}
\alias{guess_read_fnc}
\title{Determine which read function to use on data}
\usage{
guess_read_fnc(pkg_dir_full_paths)
}
\arguments{
\item{pkg_dir_name}{(character) Character vector of full paths to directories containing data and metadata}
}
\value{
(character) Named character vector of recommended read functions. Indices correspond to directory order as supplied. Names are the directory names.
}
\description{
Determine which read function to use on data
}
