% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cat_var.R
\name{plot_cat_var}
\alias{plot_cat_var}
\title{plot_cat_var}
\usage{
plot_cat_var(df, var, varname)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}

\item{varname}{(character) Name of variable of interest.}
}
\value{
A ggplot2 object containing the histogram for that column
}
\description{
Creates a histogram of the frequency with which different codes occur for categorical variables.
}
