% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_values.R
\name{plot_missing_values}
\alias{plot_missing_values}
\title{Plot missing values in a vector.}
\usage{
plot_missing_values(var)
}
\arguments{
\item{var}{(data.frame) Single-column data.frame whose missing values are to be graphed}
}
\value{
A ggplot2 graph object
}
\description{
plot_missing_values
}
\details{
Creates a plot that shows where missing values occur in the data. 
The plot appears as a bar with vertical lines representing valid or NA observations for a single variable.
}
