% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numeric_var.R
\name{plot_numeric_var}
\alias{plot_numeric_var}
\title{plot_numeric_var}
\usage{
plot_numeric_var(df, var, varname)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}

\item{varname}{(character) Name of variable of interest.}
}
\value{
A ggplot2 object containing the histogram and density plot for that column
}
\description{
Creates a histogram with density plot for numeric vectors
}
