% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_detective.R
\name{space_detective}
\alias{space_detective}
\title{the space detective}
\usage{
space_detective(entity_list, report = F)
}
\arguments{
\item{entity_list}{(list) Metajam output for a single data entity}

\item{report}{(logical) Whether the function is used in the report-making environment. Defaults to FALSE.}
}
\value{
A character vector whose two elements are named "x_col" and "y_col", indicating the columns (as appears in data) containing spatial information in the x and y dimensions respectively, or an message if not detected or if we aren't equipped to deal with the number of hits we got. If report == T, will also output messages.
}
\description{
Wrapper for \code{\link{classify_xy}}. Runs \code{\link{classify_xy}}, counts number of hits in each dimension, then decides what to depending on those two numbers.
}
