% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_plot.R
\name{space_plot}
\alias{space_plot}
\title{the Space plot}
\usage{
space_plot(entity_df, varname, space_cols = space_cols)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}

\item{varname}{Name of variable to make data availability and summary. Remember to enclose in quotes.}

\item{space_cols}{A character vector whose two elements are named "y_col" and "x_col". Outpur from \code{\link{space_detective}}}
}
\value{
For all variables, function returns a data availability plot over the study area. For categorical variables, function returns an extra plot showing most prevalent level of that variable in each spatial grid cell over the study area. For numeric variables, function returns an extra plot showing mean of that variable in each spatial grid cell over the study area. y_dim and lon are binned; there are 50 bins normalized to attempt to enforce equal scales along each dimension.
}
\description{
Create variable specific spatial heatmaps to show data availability and data summary over the geographical study area.
}
