% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_report_complete.R
\name{static_report_complete}
\alias{static_report_complete}
\title{Wrapper function to generate complete HTML report: entity-level  plus variable-level reports for one entity.}
\usage{
static_report_complete(entity_list, report_filename = NULL, output_path,
  DOI = NULL, shiny = F)
}
\arguments{
\item{entity_list}{(list) A list containing information on a single data entity in metajam output format.}

\item{report_filename}{(character) Name to give the report file. Defaults to NULL, in which case the report will be named in this  pattern "report_(file name here).html" or "report_data.html" if summary data is not available.}

\item{output_path}{(character) Path to save complete HTML report to. If "ask", a Rstudio pop-up window appears allowing choice of save directory (requires RStudio >= 1.1.287).}

\item{DOI}{(character) The data package DOI. If supplied, reports will include DOI for re-producibility.}

\item{shiny}{(logical) Whether this function is called in the GUI environment. The only difference is GUI-embedded reports do not feature a floating table of contents.}
}
\value{
A HTML file with complete report on the chosen data entity.
}
\description{
Wrapper function to generate complete HTML report: entity-level  plus variable-level reports for one entity.
}
