% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all_var.R
\name{summarize_all_var}
\alias{summarize_all_var}
\title{Summarize all variables in a data.frame}
\usage{
summarize_all_var(entity_list)
}
\arguments{
\item{entity_list}{}
}
\description{
Summarize all variables in a data.frame
}
