% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable2.R
\name{summarize_cat_levels}
\alias{summarize_cat_levels}
\title{Summarize levels of categorical variables.}
\usage{
summarize_cat_levels(entity_df)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}
}
\value{
(data.frame) A table with level-wise statistics for categorical variables.
}
\description{
Create a summary table with level-wise statistics for categorical variables.
}
