% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable3.R
\name{summarize_cat_var}
\alias{summarize_cat_var}
\title{Summarize categorical variables.}
\usage{
summarize_cat_var(entity_df)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}
}
\value{
(data.frame) A table with variable-wise statistics for categorical variables.
}
\description{
Create a summary table with variable-wise statistics for categorical variables.
}
