% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable1.R
\name{summarize_numeric_var}
\alias{summarize_numeric_var}
\title{Summarize numerical variables.}
\usage{
summarize_numeric_var(entity_df)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}
}
\value{
(data.frame) A table with variable-wise statistics for numerical variables.
}
\description{
Create a summary table with variable-wise statistics for numerical variables.
}
