% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_time_var.R
\name{summarize_time_var}
\alias{summarize_time_var}
\title{summarize_time_var}
\usage{
summarize_time_var(var)
}
\arguments{
\item{var}{(vector) A column of POSIX- or Date-type data, subsetted from the entity data frame.}
}
\value{
(data.frame) A summary table of time (POSIX, Date) objects.
}
\description{
This function writes a summary table of time (POSIX, Date) objects
}
